requirements.txt
Flask
Gunicorn
flask_cors
flask_mysqldb
python-dotenv
Flask-JWT-Extended
Werkzeug
PyMySQL
flask_mail
flask-socketio
eventlet